Initial Commit of Sample Site
Current Status: No connection to tmdb.com is being made. All Movie Data
has been stored within the app as JSON files and will be requested and
parsed by the Action Classes (everything is prepared to be able to
switch to the tmdb.com API though). There is no HTML version yet. The
Tactile version re-uses Biscuit ML files where possible. Because of one
issue with an include the actual Movie Stats are not yet re-used.